+++ CHANGELOG.md
@@ -9,15 +9,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Attributes may now also reference variables in scope, by prefixing a name with
+  `$`. There is no builtin use for this feature.
+
 ### Removed
 
 ### Changed
 
 ### Fixed
 
+* Propagation of flattening attributes in some cases.
+
+* GPU code generation now supports reductions with more than 25 operators.
+
+* `reduce` and `reduce_comm` now produce alias-free results.
+
 ## [0.27.1]
 
-## Added
+### Added
 
 * Rewrote partition to be a single kernel.
 
@@ -36,12 +45,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * A simplification rule for branches that return common results (#2526).
 
-## Changed
+### Changed
 
 * Local functions are no longer let-generalised (i.e., made polymorphic).
   Explicitly polymorphic local functions are still supported.
 
-## Fixed
+### Fixed
 
 * Filter now uses the predicate once per element instead of twice.